requirements.txt
fastapi==0.96.0
uvicorn[standard]==0.22.0
python-telegram-bot==20.4
requests==2.31.0
python-dotenv==1.0.0
